fix(pointer): preview event rate, four shader and metadata corrections, test split
The preview fed the sampler one event per frame while the compositor feeds
real pointer events at 125-1000 Hz. Since the sampler places a slot from the
gap since the last one, an append could only ever land on a 16 ms tick
boundary: the effective spacing was the interval rounded up to a multiple of
16, and the ring spanned about half again as long as the pack's window. It
now walks the path covered during the tick at 250 Hz, which is what makes
the preview's trail the same length as the one on screen.

phosphor-trail tracked hueAge against max(core, halo), two independently
accumulated maxima that between them can exceed any single segment's pair,
so past that point no segment could win and the hue stopped following the
one that decides the final coverage. Tracked against a running best of the
same quantity instead.

Three descriptions said the wrong thing. burst's releaseSpray is used for
the bit count and the launch speed, never the bit size. click-ripple's Size
promised a diameter the reach clamp pulls in at large thicknesses. sparks'
header claimed the head slot's re-roll was invisible, which was true when
the sample interval sat near its 8 ms floor and is not at 33 ms.

test_pack_validators.cpp reached 1313 lines once the pointer rules got the
coverage they were missing, past the 1150 ceiling. The pointer arm moves to
test_pointer_pack_validator.cpp; both files are now well under it and no
slot was lost.

Also: sparks' reachParam moves beside `layer` where the other ten packs keep
it, brackets loses the two trailing full stops that made it the only file
with mixed punctuation, and the afterglow buffer pass now spells out that
reordering its parameters array silently swaps what it reads by slot.

Full suite 461/461.
