+++ data/pointer/phosphor-trail/effect.frag
@@ -87,6 +87,7 @@ vec4 pPointer(vec2 uv) {
     float core = 0.0;
     float halo = 0.0;
     float hueAge = 0.0;
+    float bestCover = 0.0;
 
     // Only the LIVE run of samples is drawn or smoothed over: at a lifetime
     // equal to the metadata trailSeconds the samples behind it are outside
@@ -134,7 +135,16 @@ vec4 pPointer(vec2 uv) {
         // display as every sibling pack's.
         float c = (1.0 - smoothstep(halfWidth - 0.75, halfWidth + 0.75, d)) * fade;
         float h = exp(-(d * d) / (2.0 * sigma * sigma)) * fade * 0.45 * max(p_glow, 0.0);
-        if (max(c, h) > max(core, halo)) {
+        // Tracked against a running best of the SAME quantity `cover` is
+        // taken from below. Comparing against max(core, halo) instead used
+        // two independently accumulated maxima, which between them can exceed
+        // any single segment's max(c, h), so past that point no segment could
+        // win and hueAge stopped following the one that actually decides the
+        // colour. (The two maxima agree in value — a max of pairwise maxima is
+        // the max of the separate maxima — but only this form has an argmax.)
+        float thisCover = max(c, h);
+        if (thisCover > bestCover) {
+            bestCover = thisCover;
             hueAge = age;
         }
         core = max(core, c);
