+++ data/pointer/sparks/metadata.json
@@ -7,6 +7,7 @@
     "category": "Particles",
     "fragmentShader": "effect.frag",
     "layer": "below",
+    "reachParam": "reach",
     "trailSeconds": 1.0,
     "parameters": [
         {
@@ -104,6 +105,5 @@
             "default": "#ffa855f7",
             "description": "Colour a spark fades toward before it is gone"
         }
-    ],
-    "reachParam": "reach"
+    ]
 }
